refactor(handlers): extract context.ts handler from intent.ts (Stage B)
handleSAPContext + CDS upstream/sibling helpers. 3672 tests, A1 zero-diff.
